# Speed up VSO time parsing in some cases