--- a/changelog/5681.trivial.rst
+++ b/changelog/5681.trivial.rst
@@ -1,0 +1,2 @@
+Sped up the parsing of results from the VSO. For large queries this significantly
+reduces the time needed to perform a query to the VSO.