@dstansby 2021-11-07
Note to self when I get back to this, `Time(val=0, val2=np.nan, format='mjd')` is the best way to create a masked time at the moment in astropy.